Remove activation/deactivation from Makefile
Each line of the Makefile happens in its own shell, which is anyway a
different shell from the current one, so you can't really statefully
activate a virtual environment from a Makefile.

It seems like the thing to do is to do continuation like

https://stackoverflow.com/questions/24736146/how-to-use-virtualenv-in-makefile

so that you source the env right before you run whatever you want to run
in the env.